+++ mavlink-core/src/connection/direct_serial/mod.rs
@@ -1,0 +1,7 @@
+pub mod config;
+mod sync;
+
+#[cfg(feature = "tokio")]
+mod r#async;
+
+pub use sync::SerialConnection;
